moved code sections into separate source files
